+++ lib/micha_core.dart
@@ -1,5 +1,6 @@
 library micha_core;
 
+export 'src/async/async_builder.dart';
 export 'src/extensions/by_name_or_null.dart';
 export 'src/extensions/separated.dart';
 export 'src/extensions/times.dart';